test : 테스트코드 파일로 분리 및 오류처리 테스트 추가
